// src/app/interfaces/common/notices.interface.ts
/* eslint-disable prettier/prettier */
export interface Notices {
  _id?: string;
  name?: string;
  nameEn?: string;
  image?: string;
  seoImage?: string;
  title?: string;
  shortDescription?:string;
  description?: string;
  shortDescriptionEn?:string;
  descriptionEn?: string;
  createdAt?: Date;
  updatedAt?: Date;
}